+++ src/main/java/ca/corefacility/bioinformatics/irida/repositories/analysis/submission/AnalysisSubmissionRepository.java
@@ -1,9 +1,6 @@
 package ca.corefacility.bioinformatics.irida.repositories.analysis.submission;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.Set;
-import java.util.UUID;
+import java.util.*;
 
 import org.springframework.data.jpa.repository.Query;
 
@@ -17,32 +14,29 @@ import ca.corefacility.bioinformatics.irida.model.workflow.analysis.Analysis;
 import ca.corefacility.bioinformatics.irida.model.workflow.analysis.ProjectSampleAnalysisOutputInfo;
 import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSubmission;
 import ca.corefacility.bioinformatics.irida.repositories.IridaJpaRepository;
+import ca.corefacility.bioinformatics.irida.ria.web.admin.dto.statistics.GenericStatModel;
 
 /**
  * A repository for managing {@link AnalysisSubmission} objects.
- * 
- *
  */
 public interface AnalysisSubmissionRepository extends IridaJpaRepository<AnalysisSubmission, Long> {
 
 	/**
 	 * Loads up a list of {@link AnalysisSubmission}s with the given state.
-	 * 
-	 * @param state
-	 *            The state of the analyses to search for.
+	 *
+	 * @param state The state of the analyses to search for.
 	 * @return A {@link List} of {@link AnalysisSubmission} objects with the
-	 *         given state.
+	 * given state.
 	 */
 	@Query("select s from AnalysisSubmission s where s.analysisState = ?1")
 	public List<AnalysisSubmission> findByAnalysisState(AnalysisState state);
-	
+
 	/**
 	 * Loads up a list of {@link AnalysisSubmission}s with the given state.
-	 * 
-	 * @param state
-	 *            A collection of states to search for.
+	 *
+	 * @param state A collection of states to search for.
 	 * @return A {@link List} of {@link AnalysisSubmission} objects with the
-	 *         given state.
+	 * given state.
 	 */
 	@Query("select s from AnalysisSubmission s where s.analysisState in ?1")
 	public List<AnalysisSubmission> findByAnalysisState(Collection<AnalysisState> state);
@@ -56,41 +50,36 @@ public interface AnalysisSubmissionRepository extends IridaJpaRepository<Analysi
 	@Query("select count(s.id) from AnalysisSubmission s where s.analysisState in ?1")
 	public Long countByAnalysisState(Collection<AnalysisState> state);
 
-
 	/**
 	 * Loads up a list of {@link AnalysisSubmission}s with the given states.
-	 * 
-	 * @param analysisState
-	 *            The {@link AnalysisState} of the analyses to search for.
-	 * @param analysisCleanedState
-	 *            The {@link AnalysisCleanedState} of the analyses to search
-	 *            for.
+	 *
+	 * @param analysisState        The {@link AnalysisState} of the analyses to search for.
+	 * @param analysisCleanedState The {@link AnalysisCleanedState} of the analyses to search
+	 *                             for.
 	 * @return A {@link List} of {@link AnalysisSubmission} objects with the
-	 *         given states.
+	 * given states.
 	 */
 	@Query("select s from AnalysisSubmission s where s.analysisState = ?1 and s.analysisCleanedState = ?2")
 	public List<AnalysisSubmission> findByAnalysisState(AnalysisState analysisState,
 			AnalysisCleanedState analysisCleanedState);
-	
+
 	/**
 	 * Finds all {@link AnalysisSubmission}s corresponding to the given workflow
 	 * ids.
-	 * 
-	 * @param workflowIds
-	 *            The workflow ids to match.
+	 *
+	 * @param workflowIds The workflow ids to match.
 	 * @return A list of {@link AnalysisSubmission}s matching one of the
-	 *         workflow ids.
+	 * workflow ids.
 	 */
 	@Query("select s from AnalysisSubmission s where s.workflowId in ?1")
 	public List<AnalysisSubmission> findByWorkflowIds(Collection<UUID> workflowIds);
 
 	/**
 	 * Loads up all {@link AnalysisSubmission}s by the submitted {@link User}.
-	 * 
-	 * @param submitter
-	 *            The {@link User} who submitted the analysis.
+	 *
+	 * @param submitter The {@link User} who submitted the analysis.
 	 * @return A {@link List} of {@link AnalysisSubmission}s by the {@link User}
-	 *         .
+	 * .
 	 */
 	@Query("select s from AnalysisSubmission s where s.submitter = ?1")
 	public Set<AnalysisSubmission> findBySubmitter(User submitter);
@@ -98,9 +87,8 @@ public interface AnalysisSubmissionRepository extends IridaJpaRepository<Analysi
 	/**
 	 * Finds the {@link AnalysisSubmission} that caused the passed
 	 * {@link Analysis} to be created.
-	 * 
-	 * @param analysis
-	 *            the analysis to find the submission for
+	 *
+	 * @param analysis the analysis to find the submission for
 	 * @return the submission for the analysis
 	 */
 	@Query("select s from AnalysisSubmission s where s.analysis = ?1")
@@ -109,9 +97,8 @@ public interface AnalysisSubmissionRepository extends IridaJpaRepository<Analysi
 	/**
 	 * Get the Set of {@link AnalysisSubmission}s which use a given
 	 * {@link SequencingObject}
-	 * 
-	 * @param object
-	 *            The {@link SequencingObject} to get submissions for
+	 *
+	 * @param object The {@link SequencingObject} to get submissions for
 	 * @return Set of {@link AnalysisSubmission}
 	 */
 	@Query("FROM AnalysisSubmission s WHERE ?1 IN elements(s.inputFiles)")
@@ -120,9 +107,8 @@ public interface AnalysisSubmissionRepository extends IridaJpaRepository<Analysi
 	/**
 	 * Get the Set of {@link AnalysisSubmission}s making use of the given
 	 * {@link ReferenceFile}.
-	 * 
-	 * @param file
-	 *            The {@link ReferenceFile}.
+	 *
+	 * @param file The {@link ReferenceFile}.
 	 * @return A Set of {@link AnalysisSubmission}s.
 	 */
 	@Query("FROM AnalysisSubmission s WHERE ?1 = referenceFile")
@@ -139,19 +125,44 @@ public interface AnalysisSubmissionRepository extends IridaJpaRepository<Analysi
 	/**
 	 * Get all {@link ProjectSampleAnalysisOutputInfo} shared with a {@link Project}.
 	 *
-	 * @param projectId {@link Project} id
+	 * @param projectId   {@link Project} id
 	 * @param workflowIds Workflow UUIDs of workflow pipelines to get output files for
 	 * @return a list of {@link ProjectSampleAnalysisOutputInfo}
 	 */
-	List<ProjectSampleAnalysisOutputInfo> getAllAnalysisOutputInfoSharedWithProject(Long projectId, Set<UUID> workflowIds);
+	List<ProjectSampleAnalysisOutputInfo> getAllAnalysisOutputInfoSharedWithProject(Long projectId,
+			Set<UUID> workflowIds);
 
 	/**
 	 * Get all automated {@link ProjectSampleAnalysisOutputInfo} for a {@link Project}.
 	 *
-	 * @param projectId {@link Project} id
+	 * @param projectId   {@link Project} id
 	 * @param workflowIds Workflow UUIDs of workflow pipelines to get output files for
 	 * @return a list of {@link ProjectSampleAnalysisOutputInfo}
 	 */
 	List<ProjectSampleAnalysisOutputInfo> getAllAutomatedAnalysisOutputInfoForAProject(Long projectId,
 			Set<UUID> workflowIds);
+
+	/**
+	 * Get the count of {@link AnalysisSubmission}s run in time period
+	 *
+	 * @param createdDate The minimum created date for the analysis submission
+	 * @return A count of {@link AnalysisSubmission}s run in time period.
+	 */
+	@Query("select count(s.id) from AnalysisSubmission s where s.createdDate >= ?1")
+	public Long countAnalysesRanInTimePeriod(Date createdDate);
+
+	/**
+	 * Get a list of {@link GenericStatModel}s for analyses ran in past n time period
+	 * grouped by format provided.
+	 *
+	 * @param createdDate The minimum created date for the analysis submission
+	 * @param groupByFormat The format to use for grouping the results.
+	 * @return A list of {@link GenericStatModel}s
+	 */
+	@Query("select new ca.corefacility.bioinformatics.irida.ria.web.admin.dto.statistics.GenericStatModel(function('date_format', s.createdDate, ?2), count(s.id))"
+			+ "from AnalysisSubmission s where s.createdDate >= ?1 group by function('date_format', s.createdDate, ?2)")
+	public List<GenericStatModel> countAnalysesRanGrouped(Date createdDate, String groupByFormat);
+
 }
+
+